修改部分代码结构
