+++ src/Front-end/package.json
@@ -10,18 +10,17 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "axios": "^0.18.0",
     "babel-polyfill": "^6.26.0",
     "bootstrap": "^3.3.7",
     "css-loader": "^1.0.0",
     "echarts": "^4.1.0",
-    "element-ui": "^2.4.6",
     "file-loader": "^1.1.11",
     "html-webpack-plugin": "^3.2.0",
     "jquery": "^3.0.0",
     "lodash": "^4.17.10",
     "mint-ui": "^2.2.13",
     "style-loader": "^0.21.0",
+    "uglifyjs-webpack-plugin": "^2.1.3",
     "url": "^0.11.0",
     "url-loader": "^1.0.1",
     "vue": "^2.5.16",
