+++ src/Front-end/webpack.config.js
@@ -4,15 +4,15 @@ const htmlWebpackPlugin = require('html-webpack-plugin');
 const webpack = require("webpack");
 
 module.exports = {
-    mode: 'development',
-    // mode: 'production',
+    // mode: 'development',
+    mode: 'production',
     //指定入口
-    // entry: path.join(__dirname, "./main.js"),
+    entry: path.join(__dirname, "./main.js"),
     //使用babel-polyfill模拟ES6使用的环境，否则在IE使用vuex会出现“vuex requires a Promise polyfill in this browser”错误。
     //目前IE可支持最低版本为IE10
-    entry: {
-        app: ["babel-polyfill", path.join(__dirname, "./main.js")]
-    },
+    // entry: {
+    //     app: ["babel-polyfill", path.join(__dirname, "./main.js")]
+    // },
     //输出配置
     output: {
         path: path.join(__dirname, "../main/resources/dist/"), //指定输出目录
@@ -46,7 +46,11 @@ module.exports = {
             //引入jquery
             $: "jquery",
             jQuery: "jquery",
-            "windows.jQuery": "jquery"
+            "windows.jQuery": "jquery",
+            //引入 lodash/debounce
+            debounce: "lodash/debounce",
+            // 引入 echarts/dist/echarts
+            echarts: "echarts/dist/echarts"
         })
-    ]
-}
+    ],
+};
