+++ src/main/java/defaultpackage/Application.java
@@ -4,8 +4,12 @@ import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+import java.text.SimpleDateFormat;
+import java.time.LocalDate;
+import java.util.Date;
+
 @SpringBootApplication
-@MapperScan("defaultpackage.mybatis.mapping")
+@MapperScan("defaultpackage.mybatis.dao")
 public class Application {
 
     public static void main(String[] args) {
