+++ src/main/resources/mapper/FundDataMapping.xml
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="defaultpackage.mybatis.dao.FundDataMapping">
+    <!--获取对象集合，每个对象包括：基金代码、基金名称、基金拼音简称、基金类型-->
+    <select id="GetFundNameData" resultType="defaultpackage.mybatis.domain.EntityFundNameData">
+        SELECT fundCode, fundName, fundShortName, fundType FROM fundnamedata;
+    </select>
+
+    <!--根据基金代码获取全净值-->
+    <!--获取对象集合，每个对象包括：基金代码、基金单位净值日期、基金单位净值、基金累计净值、基金净值日增长率、基金申购状态、基金赎回状态、分红配送-->
+    <select id="GetFundValueDataByCode" resultType="defaultpackage.mybatis.domain.EntityFundValueData" statementType="STATEMENT">
+        SELECT fundCode, valueDate, NAV, cumulativeNAV, dailyGrowthRate, subscribeStatus, redemptionStatus, melonCutting
+        FROM ${tableName} WHERE fundCode=${fundCode} ORDER BY valueDate ASC;
+    </select>
+
+    <!--获取指定日期净值-->
+    <!--获取对象集合，每个对象包括：基金代码、基金单位净值日期、基金单位净值、基金累计净值、基金净值日增长率、基金申购状态、基金赎回状态、分红配送-->
+    <select id="GetFundValueDataLimit" resultType="defaultpackage.mybatis.domain.EntityFundValueData" statementType="STATEMENT">
+        SELECT fundCode, valueDate, NAV, cumulativeNAV, dailyGrowthRate, subscribeStatus, redemptionStatus, melonCutting
+        FROM ${tableName} WHERE fundCode = ${fundCode}
+        <![CDATA[
+        AND DATE_FORMAT(valueDate,'%Y-%m-%d') >= DATE_FORMAT(${startDate}, '%Y-%m-%d')
+        AND DATE_FORMAT(valueDate,'%Y-%m-%d') <= DATE_FORMAT(${endDate}, '%Y-%m-%d') ORDER BY valueDate ASC;
+        ]]>
+    </select>
+
+    <select id="GetFundPerformanceByCode" resultType="defaultpackage.mybatis.domain.EntityFundPerformance">
+        SELECT fundCode, valueDate, fundIncomeRate30Day, fundIncomeRate90Day, fundIncomeRate180Day, fundIncomeRate360Day, fundIncomeRate1080Day, fundIncomeRateCreatedToNow, fundJensenAlpha90Day, fundJensenAlpha180Day, fundJensenAlpha360Day, fundMaxDown180Day, fundMaxDownTime180Day, fundMaxDown360Day, fundMaxDownTime360Day, fundDownStdevp360Day
+        FROM fundperformance
+        WHERE fundCode=#{fundCode}
+    </select>
+
+    <!--根据基金类型，查询数据库，返回该类型下每支基金的17条基本数据：1基金代码，2基金名称，3拼音简称，4净值日期，5净值日期对应单位净值，6累计净值，7单位净值日增长率，8近1周增长率，9近1月增长率，10近3月增长率，11近6月增长率，12近1年增长率，13近2年增长率，14近3年增长率，15今年来增长率，16成立以来收益率，17成立日期-->
+    <select id="GetFundBaseDataLimit" parameterType="java.util.Map" resultType="defaultpackage.mybatis.domain.EntityFundBaseData">
+        SELECT fundCode, fundName, fundShortName, fundType, fundLastNAVDate, fundLastNAV, fundCumulativeNAV, fundNAVGrowthRate, fundIncomeRateOneWeek, fundIncomeRateOneMonth, fundIncomeRateThreeMonth, fundIncomeRateSixMonth, fundIncomeRateOneYear, fundIncomeRateTwoYear, fundIncomeRateThreeYear, fundIncomeRateCurrentYear, fundIncomeRateCreatedToNow, fundCreatedDate
+        FROM fundbasedata
+        WHERE fundType LIKE CONCAT("%",#{fundType},"%") AND ID >=(SELECT ID FROM fundbasedata WHERE fundType LIKE CONCAT("%",#{fundType},"%") ORDER BY ID LIMIT #{startIndex},1)
+        LIMIT #{dataNum};
+    </select>
+
+    <!--获取分页查询的查询结果总记录数-->
+    <select id="GetAllResultCountForPaging" parameterType="java.util.Map" resultType="Integer">
+        SELECT COUNT(ID) FROM fundbasedata WHERE fundType LIKE CONCAT("%",#{fundType},"%");
+    </select>
+
+    <!--根据查询字符串，查询符合条件的基金的17条基本数据-->
+    <select id="GetFundBaseDataBySearchStr" parameterType="java.util.Map" resultType="defaultpackage.mybatis.domain.EntityFundBaseData">
+        SELECT fundCode, fundName, fundShortName, fundType, fundLastNAVDate, fundLastNAV, fundCumulativeNAV, fundNAVGrowthRate, fundIncomeRateOneWeek, fundIncomeRateOneMonth, fundIncomeRateThreeMonth, fundIncomeRateSixMonth, fundIncomeRateOneYear, fundIncomeRateTwoYear, fundIncomeRateThreeYear, fundIncomeRateCurrentYear, fundIncomeRateCreatedToNow, fundCreatedDate
+        FROM fundbasedata
+        WHERE fundCode LIKE CONCAT("%",#{searchStr},"%") OR fundName LIKE CONCAT("%",#{searchStr},"%") OR fundShortName LIKE CONCAT("%",#{searchStr},"%");
+    </select>
+
+    <!--根据基金代码和指定时间（传入净值最新时间），查询基金目前现任的所有基金经理-->
+    <!--大于等于净值日期的筛选条件是为了将周末天数也计算在内，如：周末录入的基金经理数据，若传入周五的净值日期，则无法匹配-->
+    <select id="GetFundManagerDataByCodeAndDate" parameterType="java.util.Map" resultType="defaultpackage.mybatis.domain.EntityFundManagerData">
+        SELECT dataDate, managerCode, managerName, companyCode, companyName, currentFundCodeList, currentFundNameList, managerWorkTime, currentFundMaxIncome, currentFundMaxIncomeCode, currentFundMaxIncomeName, currentAssets, managerWorkTimeMaxIncome
+        FROM fundmanagerdata
+        WHERE currentFundCodeList LIKE CONCAT("%",#{fundCode},"%") AND dataDate>=#{fundLastNAVDate}
+    </select>
+</mapper>
